# Don't alert.UnexpectedEvent for dropped hit-tracker-client hits