@vadimberezniker 2025-06-03
This seems a bit complicated. 
At this point you might as well introduce a new metric specific to the hit tracker and configure a custom alert as you'd like.